AttributeError Solved & 11.ORM Package
AttributeError solved
ORM을 Package로 나눠서 분류 완료
Trial and error
